Created Book Service + Controller + Builder

+++ src/main/java/com/leonardo/bookstoremanager/dto/BookResponseDTO.java
@@ -15,9 +15,9 @@ public class BookResponseDTO {
 
     private String isbn;
 
-    private Long pages;
+    private Integer pages;
 
-    private Long chapters;
+    private Integer chapters;
 
     private AuthorDTO authorId;
